refactor: Refactorizar importaciones de los componentes en el Dashboard
